+++ providers/provider_default.go
@@ -10,8 +10,8 @@ import (
 
 	"github.com/coreos/go-oidc"
 
-	"github.com/oauth2-proxy/oauth2-proxy/pkg/apis/sessions"
-	"github.com/oauth2-proxy/oauth2-proxy/pkg/requests"
+	"github.com/oauth2-proxy/oauth2-proxy/v7/pkg/apis/sessions"
+	"github.com/oauth2-proxy/oauth2-proxy/v7/pkg/requests"
 )
 
 var _ Provider = (*ProviderData)(nil)